[KAIZEN-0] bytte ut useFetch for react-query for gsaktema
react-query er mer brukt, aktivt forvaltet og støtter refetching mye mer elegant.
